backend/requirements.txt
fastapi[all]==0.105.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
pymongo==4.4.1
fastapi-utilities==0.1.3.1
uvicorn[standard]==0.23.2
passlib[bcrypt]==1.7.4
bcrypt==4.1.2